chore: add external routes and organize route files
