+++ app/data/dashboardLinks/subHeaderLinks.ts
@@ -2,7 +2,7 @@ import {
   getContactsPageRoute,
   getOperatorsPageRoute,
   getProjectsPageRoute,
-} from "pageRoutes";
+} from "routes/pageRoutes";
 
 const subHeaderLinks = [
   {
