+++ app/hooks/useRedirectToContacts.tsx
@@ -1,6 +1,6 @@
 import { useEffect } from "react";
 import { useRouter } from "next/router";
-import { getContactsPageRoute } from "pageRoutes";
+import { getContactsPageRoute } from "routes/pageRoutes";
 /**
  *  Hook that redirects to the contacts list page, if the contact being viewed is committed and the `preventRedirect` flag is false.
  *  If the redirect is happening, then the hook returns true, and false otherwise.
